+++ sentry_sdk/consts.py
@@ -250,4 +250,4 @@ DEFAULT_OPTIONS = _get_default_options()
 del _get_default_options
 
 
-VERSION = "1.29.0"
+VERSION = "1.29.2"