+++ sentry_sdk/crons/decorator.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 from functools import wraps
 import sys
 
@@ -10,15 +9,16 @@ from sentry_sdk.utils import now
 
 
 if TYPE_CHECKING:
-    pass
+    from typing import Any, Callable, Optional
 
 
+@contextmanager
 def monitor(monitor_slug=None):
-    # type: (Optional[str]) -> Callable[..., Any]
+    # type: (Optional[str]) -> Generator[None, None, None]
     """
-    Decorator to capture checkin events for a monitor.
+    Decorator/context manager to capture checkin events for a monitor.
 
-    Usage:
+    Usage (as decorator):
     ```
     import sentry_sdk
 
@@ -32,44 +32,41 @@ def monitor(monitor_slug=None):
 
     This does not have to be used with Celery, but if you do use it with celery,
     put the `@sentry_sdk.monitor` decorator below Celery's `@app.task` decorator.
+
+    Usage (as context manager):
+    ```
+    import sentry_sdk
+
+    def test(arg):
+        with sentry_sdk.monitor(monitor_slug='my-fancy-slug'):
+            print(arg)
+    ```
+
+
     """
 
-    def decorate(func):
-        # type: (Callable[..., Any]) -> Callable[..., Any]
-        if not monitor_slug:
-            return func
+    start_timestamp = now()
+    check_in_id = capture_checkin(
+        monitor_slug=monitor_slug, status=MonitorStatus.IN_PROGRESS
+    )
 
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            # type: (*Any, **Any) -> Any
-            start_timestamp = now()
-            check_in_id = capture_checkin(
-                monitor_slug=monitor_slug, status=MonitorStatus.IN_PROGRESS
-            )
+    try:
+        yield
+    except Exception:
+        duration_s = now() - start_timestamp
+        capture_checkin(
+            monitor_slug=monitor_slug,
+            check_in_id=check_in_id,
+            status=MonitorStatus.ERROR,
+            duration=duration_s,
+        )
+        exc_info = sys.exc_info()
+        reraise(*exc_info)
 
-            try:
-                result = func(*args, **kwargs)
-            except Exception:
-                duration_s = now() - start_timestamp
-                capture_checkin(
-                    monitor_slug=monitor_slug,
-                    check_in_id=check_in_id,
-                    status=MonitorStatus.ERROR,
-                    duration=duration_s,
-                )
-                exc_info = sys.exc_info()
-                reraise(*exc_info)
-
-            duration_s = now() - start_timestamp
-            capture_checkin(
-                monitor_slug=monitor_slug,
-                check_in_id=check_in_id,
-                status=MonitorStatus.OK,
-                duration=duration_s,
-            )
-
-            return result
-
-        return wrapper
-
-    return decorate
+    duration_s = now() - start_timestamp
+    capture_checkin(
+        monitor_slug=monitor_slug,
+        check_in_id=check_in_id,
+        status=MonitorStatus.OK,
+        duration=duration_s,
+    )